# docs/evolution/openapi-evolution.yaml
openapi: 3.0.3
info:
  title: Moonevue APIs - Current + Evolution
  version: 0.2.0
  description: |
    Contrato consolidado com:
    1) endpoints atuais implementados (gateway/auth)
    2) endpoints alvo para a evolução (v1) por fase
servers:
  - url: http://localhost:8080
    description: Gateway
  - url: http://localhost:8081
    description: Auth
security:
  - CookieAuth: []
...
paths:
...
  /payments:
    get:
      tags: [Current Payments]
      summary: List transactions (current)
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 0 }
        - in: query
          name: size
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Transactions page
    post:
      tags: [Current Payments]
      summary: Create charge (current)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChargeRequestDTO'
      responses:
        '200':
          description: Charge created
  /payments/checkout:
    post:
      tags: [Current Payments]
      summary: Create checkout draft transaction (current)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutTransactionRequest'
      responses:
        '201':
          description: Draft created
...
  /v1/transactions/{transactionId}/charges/emit:
    post:
      tags: [Evolution Transactions]
      summary: Emit charge in provider (target)
      parameters:
        - $ref: '#/components/parameters/TransactionIdLong'
      responses:
        '200':
          description: Emitted
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: sid
  parameters:
...
    TransactionIdLong:
      in: path
      name: transactionId
      required: true
      schema: { type: integer, format: int64 }
...
  schemas:
...
    CreateCheckoutTransactionRequest:
      type: object
      required: [bankConfigurationId, amount, description, instrument]
      properties:
        bankConfigurationId: { type: integer, format: int64 }
        amount: { type: number, format: double, minimum: 0.01 }
        description: { type: string }
        instrument: { type: string }
        clientId: { type: integer, format: int64, nullable: true }
        checkoutAccessMode:
          type: string
          enum: [PUBLIC, AUTHENTICATED_CLIENT]
        pixKey: { type: string, nullable: true }
        expiresInHours: { type: integer, nullable: true }
...
    ChargeRequestDTO:
      type: object
      properties:
        bank: { type: string }
        bankConfigurationId: { type: integer, format: int64 }
        payment:
          type: object
          properties:
            instrument:
              type: string
              enum: [PIX_IMMEDIATE, PIX_DUE, BOLETO]